backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg[binary]==3.2.6
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
PyJWT==2.10.1
fpdf2==2.7.8
